# labs/electric-plant/api-data/openapi.yaml
openapi: 3.0.3
info:
  title: API de Condição de Ativos — Eletroforça Indústrias Elétricas
  description: Retorna leituras atuais, limites específicos do ativo, violações e
    status reportado para um ativo do Complexo Industrial Serrana.
  version: 1.0.0
servers:
- url: /
  description: Substitua pela URL base da API confirmada pelo facilitador
paths:
  /assets/{asset_id}/condition:
    get:
      summary: Obtém a condição de um ativo
      description: Use esta operação sempre que um ID de ativo como XFR-401 precisar
        ser classificado. Ela retorna as quatro leituras, os limites específicos do
        ativo, as violações e os problemas conhecidos.
      operationId: get_asset_condition
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          description: ID do ativo permanente (MOTOR-201, GEN-301, XFR-401, DRIVE-101, VFD-501)
          title: Asset Id
        description: ID do ativo permanente (MOTOR-201, GEN-301, XFR-401, DRIVE-101, VFD-501)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCondition'
        '404':
          description: Ativo não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AssetCondition:
      properties:
        asset_id:
          type: string
          title: Asset Id
        name:
          type: string
          title: Name
        reported_status:
          $ref: '#/components/schemas/AssetStatus'
        metrics:
          additionalProperties:
            $ref: '#/components/schemas/MetricCondition'
          type: object
          title: Metrics
        violations:
          items:
            type: string
          type: array
          title: Violations
        issues:
          items:
            type: string
          type: array
          title: Issues
      type: object
      required:
      - asset_id
      - name
      - reported_status
      - metrics
      - violations
      - issues
      title: AssetCondition
    AssetStatus:
      type: string
      enum:
      - normal
      - warning
      - critical
      title: AssetStatus
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
          description: Mensagem de erro legível por humanos.
      type: object
      required:
      - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MetricCondition:
      properties:
        value:
          type: number
          title: Value
        unit:
          type: string
          title: Unit
        min:
          type: number
          title: Min
        max:
          type: number
          title: Max
        in_range:
          type: boolean
          title: In Range
      type: object
      required:
      - value
      - unit
      - min
      - max
      - in_range
      title: MetricCondition
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError